chore(aws): update product filters for aws_data_transfer resource
The previous fix to the golden file tried to resolve double products
found for `ap-southeast-1` region. First product has `APS1` prefix (correct)
and the second one has `APS4` (incorrect one). The fix worked until
the product results order returned `APS1` records first. But it was not
guaranteed. Mostly like it's an issue on AWS side.
Introducing an additional filter with the prefix should match exactly
the correct product. It might not work for regions that don't have such
prefix in `usagetype`.