test: update golden files for aws_db_instance
This removes the `Additional backup storage` cost component from resources that have `backup_retention_period == 0`. Beforehand we were showing it for these resources and only hiding it if `backup_retention_period` was nil, but for both cases there is no backup retention period so it doesn't make sense to show the price.