fix: clean up modules before re-downloading
This fixes an issue when modules with the same source have to be redownloaded, i.e. if the manifest file is missing.
When the fetcher comes across a package it has already downloaded it copies the existing version without the .git directory.
If the manifest is missing and we have to redownload, we might be downloading into an existing directory which does not have the .git directory and go-getter fails with this.