feat(azure): Add azurerm_storage_container resource as free
It represents a container in Storage Account (which can be treated as a folder).
`azurerm_storage_account` already includes "List and Create container
operations" cost component which includes the container creation. As
Azure billing page doesn't provide the data per container and only
cumulative operations count, we can treat this resource as free.