chore: Look up products with prices when fetching prices
Some resources may have identical records in CPAPI for the same product
filters, several products are always returned and they can only be
distinguished by their prices. However the first product we pick may not
have the price due to price filter and the lookup fails. Filtering the
products with prices helps to solve that.