fix: fix polluting the OS env variables (#1587)
We were setting os.Setenv at the top level to make sure that the sync usage functionality got any AWS_* creds set in the config file's env block. However, this can cause race conditions across goroutines, so instead we pass the project's env vars through to where we load the AWS config and use a mutex when loading this.