enhance: display correct error if module detected as registry module and fails
Before we were falling back to treating the module as a remote module if there were any issues looking it up from the registry. This changes that to differentiate between if it's not a registry module or it is a registry module but there was an error looking it up.