fix: detect correct relative filepath for vcs subpath
Fixes issue where `git rev-parse --show-toplevel` was showing the /private
dir in its output. This means that the relative vcs subpath was showing an invalid
path. I've changed the get top-level functionality to use go git, which does not
have this problem and reduces our dependence on the git binary.

Also changes functionality to return an empty string if the working directory is the top
level git project. This means that `/.` is no longer appended to the project name in
output.