feat: Add --summary-only option to Bitbucket's output/comment commands
Bitbucket does not support collapsible sections in pull request comments
so the comment can become too big if the diff has more than a few changes.

This would result in the Infracost pull request comment to only have the
top summary table. This is the simplest option but it means users don't
have the diff, which is useful if the summary shows something that requires
further investigation.