fix: fix references for `aws_ecs_service` to `aws_ecs_task_definition` with the `family` attribute
This uses the `CustomRefIDFunc` functionality to create the correct reference. Since `family` isn't really an ID, I added a check in the `aws_ecs_service` resource to make sure that the returned resource data type is an `aws_ecs_task_definition`.

I'm not sure if we have a current way of locking references to an expected target type and attribute. If not, it's probably something we want to explore in the future, or work out automatically based on the HCL.