refactor: add top level vcs metadata to Infracost JSON (#1966)
* refactor: add top level vcs metadata to Infracost JSON

Changes Infracost output JSON to include a top level VCS metadata. This is then populated to Infracost Cloud under the `repoMetadata` run metadata key. With this change I've now extracted all the detect VCS metadata functionality to the `vcs` package. This is now initialised once for the `run` commands `breakdown` and `diff`.

* fix: fetch only remote data for origin

* fix: have warning output on one line