fix: don't assume role when parsing Terragrunt (#1977)
Previously if the user had `iam_role` set in their Terragrunt config we were assuming the role like Terragrunt does. This was failing in cases where the AWS credentials were not availability in the environment. Since we're parsing the HCL we don't make any calls to the AWS APIs so don't need to assume the role.