fix: Fix GitHub comment truncation when it's too long
GitHub can't accept a comment body larger than 262144 bytes. The
previous fix incorrectly used rune length of the message instead of byte
length. Rune can be 1-4 bytes, thus the discrepancy between the values.
As the truncateMiddle method uses rune length the q value is used to
convert the length to rune one based on the message.