fix: module source path
fixes issue where module exclusion was not working as the path was not relative to the working directory.
This meant that modules like `module/t` was being marked as `../module/t`. We now join the path to the working directory
evaluating `..module/t` under the `./dev` path would result in `module/t`