fix: container_cluster setting nil node pool as sub resource (#2184)
* fix: container_cluster setting nil node pool as sub resource

fixes: https://github.com/infracost/infracost/issues/2180

With the changes in `0.10.15` we have introduced a helper function `computeCostComponents` to help with custom instance types. However this now returns an error and alters `ContainerNodePool` to return nil on unsupported machine types. See https://github.com/infracost/infracost/blob/master/internal/resources/google/container_node_pool.go#L48-L52. The `ContainerCluster` resource was incorrectly setting these nils as valid sub resources which caused a panic.

Adds an additional pipeline step to build and upload release assets to the PR so that we can send these to contributors to review and see if it fixes there issue.

* fix: golden file tests