fix: panic arising from cty conversion of cty.NilVal
gocty.ToCtyValue panics if any type provided to it is nil. This can happen if we use NilVal at any point in the context evaluation. I've removed all references of NilVal for attributes/variable evaluation in favour of DynamicVal which is an Unknown type, and should be conversion safe. I've also wrapped the problem function in a recover statement to catch any panics that might arise due to unforseen evaluation problems. This means that the terragrunt inputs would use a safe default of mocked values rather than just exiting.