refactor: cache `block.Values` output
After profiling the infracost binary with go tool pprof it showed that Block.Values was a particularly heavy call, which is executed a number of times throughout module evaluation. Block.Values evaluates every one of the Block attributes using the EvaluationContext. This change caches the Block attribute values when they have been called > 1 times and the evaluated values don't change. On further calls to Block.Values we return this cached value if the EvaluationContext is still the same as the last call to Block.Values.

This significantly improves HCL evaluation, tests locally and in the cluster saw around 30% - 40% improvement in parsing time.