refactor: scope errors by project (#2225)
* refactor: scope errors by project

Changes the Terraform and Terragrunt providers to scope evaluation and execution errors to projects, rather than failing the whole run. We now set a metadata field on the project to indicate if the project encountered errors and then continue to evaluate other detected or configured projects. Once the run has finished and piped to an output we indicate that certain projects have failed, and display errors.

This change means that pipeline runs won't fail in a multi-project setting if only a select few projects have failed. If all project error, we bubble the errors up the stack and fail the cmd.

* fix: newline within else block

* fix: remove unecessary todo

* fix: show errored projects in diff output

* fix: remove errored projects from skipped