fix: change project paths to absolute path if initial --config-file path is abs
fixes #2231

With the change to the top-level .infracost folder, if an absolute path is used for config-file or path resolution then we fail to build the correct paths for modules. This PR changes the project path passed to the `Parser` to be abs if the initial path (--config-file/--path) is absolute.