feat: add policy resource schema
Changes underlying scan functionality to send a list of scannable resources instead of a entire breakdown. This makes the policy API call much less heavy and speeds up runs. This PR also changes some of the scannable resources (S3 Bucket, ECR, e.t.c) to include some additional references so that these are available to resolve in the policy API.