feat: support for recovery_services_vault and backup_protected_vm
This as been a pain and I've rewritten this more times than I can count as there were lots of avenues I went down that didn't work or were actually incorrect. The end result is pretty simplistic, but other routes I tried, turned out to be wildly different from pricing page or just didn't encapsulate the right things.

I'm at the point where this lines up with everything and it covers the resource, however I need a fresh pair of eyes on it as I'm unsure what to add/keep, and want to draw a line in the sand as i've spent way too much time on this.