fix: replace `RunContext` and `ProjectContext` `contextValues` with concurrency safe `ContextValues`
This change introduces the ContextValues struct to both RunContext and ProjectContext, replacing the previously used map[string]interface{}. This change is introduced to enhance thread safety in our codebase by preventing concurrent read/write operations to these shared resources, which could lead to unpredictable behavior or data races.

Prior to this change the `ContextValues()` could hit a conccurent write panic because it was returning the actual map and not a copy of it, so when we read from the map it could potentially be written to in a separate goroutine.

To fix this `Values` now returns a copy of the underlying map, and I've introduced a `GetValue` method to safely read values.