feat: add support for data.aws_default_tags block
Adds support for the `data.aws_default_tags` which returns the provider default tags.
This PR also adds support to save the infracost generated plan JSON to the local filesytem.
This aids debugging, in particular seeing what final values are for the evaluated blocks.

Work on this PR exposed a serious limitation to the HCL evaluator. See this issue https://github.com/infracost/infracost/issues/2596
for more info. The code changes in this PR go some way to solving this issue but we need more work
to fully solve it.