feat: pricing client cache
Introduces an persisent and in memory query cache for the pricing api. As part of this
change the PricingApiClient is now shared across the application codebase so that we maintain
a consistent cache used across goroutines. Furthermore, we now dedupe pricing api queries sent
to the pricing API, further reducing request size made by the CLI.

Caching functionality can be controlled by the new configuration parameters:

* `INFRACOST_PRICING_CACHE_DISABLED` which, if set to true, completely removes caching functionality
* `INFRACOST_PRICING_CACHE_OBJECT_SIZE` which can control the size of the cache. This is useful for
memory requirements.