chore: add parser benchmarks (#2660)
Adds a simple benchmark test which is run against some heavy evaluation modules.
The benchmarks are now run as part of the GO pipeline with specific mem and cpu profiles uploaded as
release assets. The addition of benchmark tests are a necessary first step to improving the performance
of the parser. Whilst these tests are still quite naive, they give us some understanding as to the positive/negative
performance impacts of changes made to the parser. Longer term we would want to assert that changes made to the evaluator
don't alter the benchmarks by a certain delta. However, this can wait for now.