fix: fix handling of unknown objects
This fixes the case if the value has already been mocked, but then we discover that it is an object.
For example, we might have the following:
  ```
  val1 = data.my_resource
  val2 = data.my_resource["index"]
  ```
When val1 is evaluated we mock `data.my_resource` to a mock string value. When val2 is evaluated we now know it is an object,
so we need to make sure we can handle this case.