fix: remove child projects from diff output if parent path errors (#2743)
introduces logic in `CompareTo` to filter out prior projects that appear as a child path of a current errored project.
This is done as sometimes errored Terragrunt evaluation returns the master path of the project
rather than the actual Terragrunt projects. This happens when Terragrunt is unable to build
a "stack" configuration (the Terragrunt project tree) because of a parsing error.

For example if we have the following tree:
```
 .
 └── infra/

        ├── dev/
        │   └── terragrunt.hcl
        └── prod/
            └── terragrunt.hcl

```

A valid `breakdown --path infra` will return projects `infra/dev` and `infra/prod`.
However, in an errored "stack" state Terragrunt will return a single project at `infra`.
We need to find all the child projects of the parent so that we can properly exclude them from the output.
Otherwise, these projects will show as "removed" and have an invalid cost decrease.