chore: add remote module calls to project metadata (#2749)
Changes `Evaluator` logic so that when a module is evaluated we set a `SourceURL` on the definition. This
is populated from the ModuleMetadata. The url is either built from the Metadata source or a newly added `DownloadUrl` which
is found by calling a registry disco to try and resolve the original source.

The new source url is then propagated to the plan JSON and eventually aggregated into a list on the `Project` `Metadata`.