fix: support dynamic iterator attribute (#2760)
This PR adds support for expanded Dynamic block child context support.
Normally, in the context of a for_each expansion we set the child
context to have a prefix of "each". This allows correct evaluation of attributes
referencing "each.value." or "each.key.". However, dynamic blocks are special cases.
Not only can they use "each" in the content block, but they also are able to
refer to the dynamic block label, e.g:

```
 dynamic "setting" {
   for_each = var.settings
   content {
     namespace = setting.value.namespace.
     ...
   }
 }
```

additionally they have an "iterator" block:
https://developer.hashicorp.com/terraform/language/expressions/dynamic-blocks
which allow the user to specify a different prefix for a for_each expansion.
For example:

```
  dynamic "foo" {
   for_each = {
      ...
   }
   iterator = device
   content {
     name = device.value.name
     ...
   }
 }
```

To add support for this logic, we now set an iteratorName on the child
Context in for each expansion if the block type is "dynamic".