refactor: skip JSON parser for Graph evaluator
This changes the HCL Graph evaluator logic to build resources directly from
the graph output. When the graph evaluator finishes visiting a node it is
assumed that we have all the relevant data/values to build a schema resource.
This means that we can essentially skip the marshalling to Plan JSON and instead
just change the resources to their CoreResource counterparts.