refactor: force https module get logic (#2789)
Changes Terraform and Terragrunt getter logic to force a https
`Get` before attempting a ssh `Get`. The Terraform logic now uses
a custom `GitGetter` which transforms the URL before `Get`. The
Terragrunt provider cannot use this as the `Getters` are hidden below
in the lib, so we try and transform the `sourceUrl` before Download.