feat: add generate auto-detect golden files (#2792)
* fix: include `terraform.tfvars` as "auto" file (#2790)

Fixes project auto-detection so we do not treat `terraform.tfvars` as
an environment var file, but instead, correctly mark them as
auto-applied to the project.

* feat: add generate auto-detect golden files

Adds auto-detect golden file logic to simplify writing `generate` tests.
This changes the `TestGenerateConfigAutoDetect` function to look at the
`./testdata/generate` directory and run a sub test for every child directory.
I've added a README in that directory to document how to write these tests.

* enhance: support . and comments in tree.txt files

* fix: show nested tests as failing

* chore: add autodetect test cases

---------

Co-authored-by: Alistair Scott <ali.scott@gmail.com>