refactor: use different mocking logic for graph evaluator IC-608
This PR changes the mocking logic in the graph evaluator to an expression
based approach. Now instead of changing the global Context we instead traverse
any expression that is found with a Diagnostic. We walk this expression tree
until we hit an expression that is a candidate for mocking. These are:

* hclsyntax.AnonSymbolExpr:
* hclsyntax.LiteralValueExpr
* hclsyntax.ScopeTraversalExpr:

and occasionally hclsyntax.FunctionCallExpr

These are Expressions that likely correspond to "incomplete" values that infracost
cannot build. E.g. hclsyntax.ScopeTraversalExpr directly relates to a "missing attribute"
or "invalid attribute" diagnostics that are 95% of the cases that we handle. These
come about through lack of complete information because attributes/blocks which are
populated from the cloud, e.g. data blocks.

Whilst this method works for the normal evaluator, i've only enabled it for the
graph evaluator at the moment as this new functionality creates occasional issues
with count/for_each attributes that rely on values not already populated in the
context - e.g. for_each attributes that rely on module outputs.