feat: add support for directory named envs (#2847)
Changes `ProjectLocator` functionality to use env namings that are defined
in directories that hold var files. Each var file is compared against a set
of possible env var names that are built from the file name and any parent
directories that hold the var files. These parent directories can only be
directories that only hold tfvar files. For example

```
.
└── infra
    └── components
        ├── foo
        │   └── main.tf
        └── variables
            └── env
                └── test
                    └── foo.tfvars
```

`foo.tfvars` would have the following directories as "candidates" for env
naming "test", "env", "variables". The list stops at "components" as it has
a child directory "foo" that contains a project.

In order to accomodate this change I've had to reorder how we "detect"
env vars into the `ProjectLocator`, introducing a new `EnvFileMatcher`
struct that is passed around the locator and to `RootPath`s. This means
that we can use the env file matching logic in the locator tree and
determine env names correctly.