feat: add skip_auto_detection for autogenerated config files
Adds a new project config key `skip_auto_detection` that is added by default
for autogenerated config files. This fixes issues with duplicate projects
being shown when a Terraform project has a child directory with another
Terraform project.

I've also removed the `terraform_workspace` output from the generated config
as this is not required and is just noise.