refactor: group diagnostic codes (#2871)
* refactor: group diagnostic codes

This change groups the diagnostic codes projects into corresponding codes.
This allows us to abstract a diagnostic into a higher order grouping of errors
that is easier to display to the user.

With this change I've removed the `Warning` construct and instead the only warning
"missing vars" into diagnostics. Adding the corresponding fields `IsError` and `FriendlyMessage`
to accomodate this merge. This means we now only have one place to define diagnostics and
codes/numbers do not collide.

Note: this change depends on various data migrations and changes we need to make in
Infracost cloud before it is safe to release.

* fix: review comments