fix: fix cyclical dependency issue when an alias provider is used
A cyclical dependency was possible in the graph evaluator. We were adding edges for every data and resource vertex to their provider. For cases where the provider was aliased in the parent module and then implicitly passed to a child module, instead of adding an edge to the aliased provider, we were adding an edge to the root provider.

This fixes that by decoding the providers blocks before creating the edges so we can keep track of all aliased providers and add the correct edges.