fix: aft module providers do not evaluate (#2837)
* test: add test case for aft module

* test: revert the missing provider handling so the errors panic

* test: refactor breakdown with tags tests

* fix: provider inheritance

* fix: don't inherit providers when populating graph

* fix: compare string value, not pointers, when checking if aliases match

* Revert "test: revert the missing provider handling so the errors panic"

This reverts commit 734cd567bfacecc4be99bae8e94edf015e12cf00.

* test: filter out inconsistent mock values

* refactor: make fmt

* fix: fix cyclical dependency issue when an alias provider is used (#2897)

A cyclical dependency was possible in the graph evaluator. We were adding edges for every data and resource vertex to their provider. For cases where the provider was aliased in the parent module and then implicitly passed to a child module, instead of adding an edge to the aliased provider, we were adding an edge to the root provider.

This fixes that by decoding the providers blocks before creating the edges so we can keep track of all aliased providers and add the correct edges.

---------

Co-authored-by: Alistair Scott <ali.scott@gmail.com>