feat: exit early if Infracost API Key is invalid (#2927)
* feat: exit early if Infracost API is invalid

Changes the `breakdown`, `diff`, `upload` and `comment` commands to exit early if
the INFRACOST_API_KEY is invalid. This is done so that users can better debug their
pipelines where an API key might be incorrectly set.

To accomodate this change i've introduced the idea of command middleware which will
wrap commands before their execution. This means we can easily chain on additional
functionality/checks in the future if required.

* fix: disable pricing cache for test

* test: only run invalid api key test

* debug

* comment out go ci stuff

* switch to new client

* fix: go.yml

* fix: use stub server for test