feat: add eks extended support cost
Changes the eks resource to use the extended support cost for the
management cost of the cluster if the version specified in the resource
is past the AWS support date. Extended support is charged at $0.6 per hour
per cluster and replaces old management cost of $0.1.

We currently hard code the cost ($0.6) as this cost is not currently available
in the pricing API.