feat: autodetect add `force_project_type` configuration
Adds the ability to configure autodetect logic to force a project to a
certain type. This is useful in cases where a directory has been detected
as two or more project types.

This configuration is meant as a fallback to fix issues where autodetect
cannot infer the correct project run configuration. Usage of this configuration
is normally idicative that we need to better handle an autodetect scenario; either by
deeper evaluation of the Terragrunt/Terraform configuration schema or some
other tree based matching logic.

Yet, in the meantime this configuration option allows the team/users to get
a working solution whilst improved logic is being worked on.