fix: filter empty path error from diff baseline
Changes the combined logic so that we filter any empty path errors that
originate from the baseline of a current project. This is done because in
the majority of cases an empty path error on the baseline just means that
the project was simply added.

This resolves a UI error in Infracost Cloud where newly added projects
were not shown because of the basline error.