feat: add support for `env_var_extensions` autodetect configuration (#2949)
* feat: add support for `env_var_extensions` autodetect configuration

Some projects contain env vars that are have non-standard naming conventions
and don't follow the common `.tfvars` and `.tfvars.json` patterns. In this
case autodetect fails to pick up the variable files as part of the `ProjectLocator`
configuration.

This change adds support for a autodetect configuration to allow us to specify
non-standard naming conventions. This, alongside the existing `env_names` configuration
field should satifsy most tfvar requirements.

However, we still have the issue of the `tfvars.json` usecase which is used by some
customers. I've had to make exceptions for this in this PR because I think adding
"configurable" support for this might be a bit more complex. In future we might
look to add a `env_var_pattern` config, which accepts a regex. This means we
could catch the prefix/suffix combo scenario.

* refactor: only use the extension matching method for custom var extensions

Only use the exact extension matching for when the custom extensions have been
specified. This way we can use the more flexible strings.HasSuffix method to
match other terraform files. This is more backwards compatible.

* refactor: rename `env_var_extensions` -> `terraform_var_file_extensions`

This makes the configuration naming more consistent with existing flags/config
`terraform_var_files`