fix: skip tfvar file decoding if name follows default conventions (#2951)
Adds a check for default tfvar naming conventions if `terraform_var_file_extensions` are specified.
This fixes an issue if the `terraform_var_file_names` contained a `.tfvars.json` extension. This
would fail the `ParseHCLFile` check and drop the autodetected var file.

A side effect of this change is that we no longer need to specify the default var file extensions
in the autodetect config, just the exception cases.