feat: add `Env` to `DetectedProject` (#2952)
This change adds a `Env` property to the `DetectedProject` struct. This allows
users to access the Infracost autodetected `env` in the projects loop. This enables
optional logic in the template. For example a common pattern is users provide environment
as the `env` as a tfvar. Additionally regions/other logic can be configured with templ
if/else logic.