fix: unknown value issues with `formatdate` and invalid dates (#2978)
This change resolves issue where usage of `formatdate` within an attribute value
caused the value to resolve to a unknown dynamic value. This was especially
problematic when used within maps and map funcitons as it caused the whole map
to return as a unknown value.

I've fixed this issue by wrapping the formatdate function, checking the return
value does not have any errors returned. If it does we return a safe default.
I've also added support for the `time_static` resource, which initially caused
us to stumble on this bug.