fix(autodetect): only use directories that contain var files when associating children (#3000)
Resolves issues with autodetect where we fail to properly associate child Terraform
var files if the root directory also has children that contain root directories. Take
the following example:

```
.
└── terraform
    ├── config
    │   ├── dev
    │   │   └── terraform.tfvars
    │   └── prod
    │       └── terraform.tfvars
    ├── main.tf
    └── foo
        ├── main.tf
        └── config
            ├── dev
            │   └── terraform.tfvars
            └── prod
                └── terraform.tfvars
```

In this case, prior to this change, when we called `AssociateChildVarFiles` on the `terraform`
directory we stopped traversing at the depth of the `foo` directory. This meant that we were
not able to associate `config/dev/terraform.tfvars` to the project.

I've now made a fix to add a method `ChildTerraformVarFiles` which traverses the tree only returning
directories that contain only tfvar files. This is safe to use in the `AssociateChildVarFiles` context
as this is only concerned with traversing to find child var files.