refactor: log output  (#3003)
* refactor: remove spinner functionality

Cleans out spinner functionality as we move to a more consistent log based approach.
Lot's of old spinners have been demoted to `Debug` level log lines as we want to keep
the `Info` level relatively clean of noise.

* refactor: log output

Refactors the progress/log output for the evaluation/project evaluation.
Chnages the output of status lines so that they are simpler and more consistent
across CLI and CI.

* All fmt.Prints are now piped to the logs, this includes all the previous ui
  package interactions, appart from a few fatal errors.
* Some Warn statements are now changed to debug
* By default the log level is INFO and is not piped to io.Discard if not explicitly
  set.
* The log formatter now checks if the env is CI and formats the log accordingly
* Local logs have their timestamps omitted and additional colours

* fix: change TFC warning log to debug

Changes the remote variable log line to be debug level rather than warn.

* fix: switch finished/starting log lines to debug

* refactor: aggregate warning logs for no price resources (#3024)

* refactor: aggregate warning logs for no price resources

Changes the prices not found warnings so they are aggregated per run. This is
deduped by resource type. In order to do this I've added a `parent` field to the
schema `Resource` so that we can find the correct resource type from a sub resource,
e.g. ebs_block -> aws_instance.

The warning messaging changes based on log level, with debug level turned on we
also augment the mesage with the resource addresses which are affected.

In addition to the warning logging I've changed the table output to add `not found`
to the components that have prices missing. This means users can see easily see wich
cost components have no prices because of pricing lookups.

* feat: add noPrice warnings to tracked env

Removes `resourceWarnings` env in favour of `pricesNotFound` env key. The former
was a nested map of counts and was pretty much impossible to use for our reporting
stack. Additionally much of the warnings, e.g. "multiple prices" were noise. Removing
gives clarity on what resource warnings we actually care about.

The new warning key is scoped by resource type and cost component name, meaning we
can scope reporting to individual cost components, which is more useful for bug triaging
and fixing.

* fix: relgate some technical warning lines to debug so that they do not pollute user

* refactor: sort no price warnings by resource count

* fix: change autodetect messaging if all projects are invalid

* fix: add `PriceFetcher` to aggregate price warnings

Changes the NoPrices functionality to instead work off a master `PriceFetcher`
struct which manages all the pricing API interactions.

* fix: update aws/azure tests following warning changes