feat: readd templatefile support (#3041)
Changes `isPathInRepo` to use wd directory instead of `RepoPath` which is likely incorrect for
Terragrunt projects.

The `RepoPath` property that is infered from detected projects was returning full project
path for detected projects when the autdetected project was inferred from a config file. A
fix for this will be addressed in a further PR.

To fix the templatefile logic I've opted to use the os working directory instead, as this is more
robust.