fix: use regex to handle version matching instead of awk (#3058)
Some release assets ended up with version 'Infracost v0.10.35+v0.10.35-v0.10.35' after the last release.  I wasn't able to reproduce the problem, but this change updates the version script to be more robust. It uses regex (instead of awk) to parse the expected tag format and falls back to using the full tag string as the version if it is in an unexpected format.